+++ InventoryTools/Logic/Columns/TypeColumn.cs
@@ -22,7 +22,7 @@ namespace InventoryTools.Logic.Columns
         public override List<MessageBase>? Draw(FilterConfiguration configuration, ColumnConfiguration columnConfiguration, SearchResult searchResult,
             int rowIndex, int columnIndex)
         {
-            if (searchResult.CuratedItem != null && (searchResult.Item.Base.IsCollectable || searchResult.Item.Base.CanBeHq))
+            if (searchResult.CuratedItem != null && (searchResult.Item.IsCollectable || searchResult.Item.Base.CanBeHq))
             {
                 ImGui.TableNextColumn();
                 if (!ImGui.TableGetColumnFlags().HasFlag(ImGuiTableColumnFlags.IsEnabled)) return null;