+++ InventoryTools/Logic/Filters/IsCollectableFilter.cs
@@ -9,7 +9,7 @@ namespace InventoryTools.Logic.Filters;
 
 public class IsCollectableFilter : GenericBooleanFilter
 {
-    public IsCollectableFilter(ILogger<IsCollectableFilter> logger, ImGuiService imGuiService) : base("IsCollectable", "Is Collectable?", "Is the item collectable?", FilterCategory.Basic, item => item.Item.Base.IsCollectable, item => item.Base.IsCollectable, logger, imGuiService)
+    public IsCollectableFilter(ILogger<IsCollectableFilter> logger, ImGuiService imGuiService) : base("IsCollectable", "Is Collectable?", "Is the item collectable?", FilterCategory.Basic, item => item.Item.IsCollectable, item => item.IsCollectable, logger, imGuiService)
     {
     }
 }